Create doi_10.21223_HMMWZH.R
This is a dataset that includes Yield, Disease Severity, and Ground Cover (LAI) throughout the potato growing season.
There are not variables for Ground Cover (I used LAI).
I've also added the weather data (this is great in CaRBO). There's  need to add the term of Hours of High Humidity (HR>90).